Create docs_links file with all documentation links
