+++ libs/ui-lib/lib/common/config/index.ts
@@ -1,1 +1,2 @@
 export * from './constants';
+export * from './docs_links';
